.history/requirement_20230404195012.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
comet-ml==3.32.6
configobj==5.0.8
contourpy==1.0.7
cycler==0.11.0
dulwich==0.21.3
everett==3.1.0
fonttools==4.39.3
idna==3.4
importlib-resources==5.12.0
jsonschema==4.17.3
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
pefile==2023.2.7
pillow==9.4.0
pip==20.1.1
pkgutil-resolve-name==1.3.10
psutil==5.9.4
pyparsing==3.0.9
pyrsistent==0.19.3
python-box==6.1.0
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0
requests-toolbelt==0.10.1
requests==2.28.2
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.10.0
sentry-sdk==1.18.0
setuptools==47.1.0
simplejson==3.18.4
six==1.16.0
thop==0.1.1.post2209072238
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tqdm==4.65.0
typing-extensions==4.5.0
ultralytics==8.0.59
urllib3==1.26.15
websocket-client==1.3.3
wrapt==1.15.0
wurlitzer==3.0.3
zipp==3.15.0